requirements.txt
customtkinter>=5.2.2
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
pyinstaller>=6.0.0
